+++ fixtures/widgets/CheckboxWidget/default.js
