+++ fixtures/widgets/RadioWidget/default.js
@@ -1,0 +1,8 @@
+module.exports = {
+  schema: {
+    type: "string",
+    title: "foo",
+    default: "b"
+  },
+  options: ["a", "b"]
+};
